Update README.md
God of War game of war is when you realize you can go anywhere. The opening setup is clear, the game goes backwards, presents you with a map of the world and says, 'You know what? You're a god, you have a boat: sort yourself out. ' The ability to go anywhere and do something in a semi-open world suddenly makes it as much about your journey through the story of Kratos as one of his goals.